move GameOverModal to folder and add CSS Module
