+++ src/components/GameOverModal/GameOverModal.tsx
@@ -1,8 +1,9 @@
 import React from 'react';
-import Modal from './Modal/Modal';
-import Button from './Button/Button';
-import { useResults } from '../hooks/useResults';
-import type { AnswerHistory } from '../types';
+import Modal from '../Modal/Modal';
+import Button from '../Button/Button';
+import { useResults } from '../../hooks/useResults';
+import type { AnswerHistory } from '../../types';
+import styles from './GameOverModal.module.css';
 
 interface GameOverModalProps {
   isOpen: boolean;
@@ -40,30 +41,30 @@ const GameOverModal: React.FC<GameOverModalProps> = ({
 
   return (
     <Modal isOpen={isOpen} onClose={onClose} title="Гра завершена!">
-      <div className="game-over-modal">
-        <div className="result-summary">
-          <div className="final-score">
-            <span className="score-value">{score}</span>
-            <span className="score-label">балів</span>
+      <div className={styles.modal}>
+        <div className={styles.summary}>
+          <div className={styles.finalScore}>
+            <span className={styles.scoreValue}>{score}</span>
+            <span className={styles.scoreLabel}>балів</span>
           </div>
           
-          <div className="result-message">
+          <div className={styles.message}>
             <p>{getResultMessage()}</p>
           </div>
 
-          <div className="detailed-stats">
-            <div className="stat-row">
+          <div className={styles.stats}>
+            <div className={styles.statRow}>
               <span>Правильних відповідей:</span>
               <strong>{correct} / {total}</strong>
             </div>
-            <div className="stat-row">
+            <div className={styles.statRow}>
               <span>Успішність:</span>
               <strong>{percent}%</strong>
             </div>
           </div>
         </div>
 
-        <div className="modal-actions">
+        <div className={styles.actions}>
           <Button onClick={onRestart} variant="primary">
             Грати знову
           </Button>
