+++ src/pages/Game/GamePage.tsx
@@ -7,7 +7,7 @@ import Question from '../../game/Question/Question';
 import AnswerList from '../../game/AnswerList/AnswerList';
 import ProgressInfo from '../../game/ProgressInfo/ProgressInfo';
 import ScoreBoard from '../../game/ScoreBoard/ScoreBoard';
-import GameOverModal from '../../components/GameOverModal';
+import GameOverModal from '../../components/GameOverModal/GameOverModal';
 import { useGameFlow } from '../../hooks/useGameFlow';
 import { useTimer } from '../../hooks/useTimer';
 import type { Question as QuestionType } from '../../types';
